cli: move cloud provider code to packages
Makes it easier to reuse code from destroy command in the upgrade command.
